chore(submission-format): modify existing solutions to match single file format
remote the two part revision files and move to the submission.js and problem-name.test.js format
