+++ package.json
@@ -10,6 +10,11 @@
   "author": "manzoid <manzoid@gmail.com>",
   "license": "MIT",
   "devDependencies": {
+    "eslint": "4.16.0",
+    "eslint-config-airbnb": "16.1.0",
+    "eslint-plugin-import": "2.8.0",
+    "eslint-plugin-jsx-a11y": "6.0.3",
+    "eslint-plugin-react": "7.5.1",
     "jest": "^22.0.6"
   }
 }
