--- single-function/add-integer-lists/add-integer-lists.js
@@ -1,1 +1,0 @@
-// Begin your set-up revision for "addTwoNumbers()" here...
