--- single-function/get-target-sum-paths/get-target-sum-paths.js
@@ -1,1 +1,0 @@
-// Begin your set-up revision for "getTargetSumPaths()" here...
