--- single-function/has-path-sum/has-path-sum.js
@@ -1,1 +1,0 @@
-// Begin your set-up revision for "hasPathSum()" here...
