--- single-function/longest-common-prefix/02-impl-revision.js
@@ -1,14 +1,0 @@
-const longestCommonPrefix = strs => {
-  if (strs.length === 0) return '';
-
-  let prefix = strs[0];
-
-  for (let i = 1; i < strs.length; i++) {
-    while (strs[i].indexOf(prefix) !== 0) {
-      prefix = prefix.substring(0, prefix.length - 1);
-    }
-  }
-  return prefix;
-};
-
-module.exports = longestCommonPrefix;
