+++ single-function/longest-common-prefix/longest-common-prefix.test.js
@@ -1,21 +1,21 @@
-const longestCommonPrefix = require('./02-impl-revision');
+const longestCommonPrefix = require('./submission.js');
 
 describe('longestCommonPrefix', () => {
   it('given empty array, [] return ""', () => {
     const arrayOfStrings = [];
-    const expected = '';
+    const expected = "";
     expect(longestCommonPrefix(arrayOfStrings)).toEqual(expected);
   });
 
   it('given array, ["aaabb", "aabb", "xaacc"] return ""', () => {
     const arrayOfStrings = ["aaabb", "aabb", "xaacc"];
-    const expected = '';
+    const expected = "";
     expect(longestCommonPrefix(arrayOfStrings)).toEqual(expected);
   });
 
-  it('given array, ["aaabb", "aabb", "aacc"] return "aa"', () => {
-    const arrayOfStrings = ["aaabb", "aabb", "aacc"];
-    const expected = 'aa';
+  it('given array, ["qqaaabb", "qqqqqqaabb", "qqqaacc"] return "aa"', () => {
+    const arrayOfStrings = ["qqaaabb", "qqqqqqaabb", "qqqaacc"];
+    const expected = "qq";
     expect(longestCommonPrefix(arrayOfStrings)).toEqual(expected);
   });
 });
